switch from `walkdir` to `jwalk` for significant gains
a full directory traversal (i.e. during `status` or `checkout`) takes
20ms less than before. `bit status` on libgit2 down to 60ms from 80ms
total.

+++ libbit/src/commit.rs
@@ -73,9 +73,9 @@ pub struct CommitSummary<'rcx> {
 impl Display for CommitSummary<'_> {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         if self.sym == SymbolicRef::HEAD {
-            writeln!(f, "[detached HEAD {}]", self.commit.oid())?;
+            writeln!(f, "[detached HEAD {}]", self.commit.oid().short())?;
         } else {
-            writeln!(f, "[{} {}]", self.sym.short(), self.commit.oid())?;
+            writeln!(f, "[{} {}]", self.sym.short(), self.commit.oid().short())?;
         }
 
         // TODO show full diffstat summary (deletions, insertions)
